+++ firehydrant/services.go
@@ -2,8 +2,6 @@ package firehydrant
 
 import (
 	"context"
-	"fmt"
-
 	"github.com/dghubble/sling"
 	"github.com/pkg/errors"
 )
@@ -29,61 +27,79 @@ func (c *RESTServicesClient) restClient() *sling.Sling {
 }
 
 // Get retrieves a service from the FireHydrant API
-// TODO: Check failure case
 func (c *RESTServicesClient) Get(ctx context.Context, id string) (*ServiceResponse, error) {
-	res := &ServiceResponse{}
-	resp, err := c.restClient().Get("services/"+id).Receive(res, nil)
+	serviceResponse := &ServiceResponse{}
+	response, err := c.restClient().Get("services/"+id).Receive(serviceResponse, nil)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not get service")
 	}
 
-	if resp.StatusCode == 404 {
-		return nil, NotFound(fmt.Sprintf("Could not find service with ID %s", id))
+	err = checkResponseStatusCode(response)
+	if err != nil {
+		return nil, err
 	}
 
-	return res, nil
+	return serviceResponse, nil
 }
 
 // List retrieves a list of services based on a service query
 func (c *RESTServicesClient) List(ctx context.Context, req *ServiceQuery) (*ServicesResponse, error) {
-	res := &ServicesResponse{}
-	_, err := c.restClient().Get("services").QueryStruct(req).Receive(res, nil)
+	servicesResponse := &ServicesResponse{}
+	response, err := c.restClient().Get("services").QueryStruct(req).Receive(servicesResponse, nil)
 	if err != nil {
-		return nil, errors.Wrap(err, "could not get service")
+		return nil, errors.Wrap(err, "could not get services")
 	}
 
-	return res, nil
+	err = checkResponseStatusCode(response)
+	if err != nil {
+		return nil, err
+	}
+
+	return servicesResponse, nil
 }
 
 // Create creates a brand spankin new service in FireHydrant
-// TODO: Check failure case
 func (c *RESTServicesClient) Create(ctx context.Context, createReq CreateServiceRequest) (*ServiceResponse, error) {
-	res := &ServiceResponse{}
-
-	if _, err := c.restClient().Post("services").BodyJSON(&createReq).Receive(res, nil); err != nil {
+	serviceResponse := &ServiceResponse{}
+	response, err := c.restClient().Post("services").BodyJSON(&createReq).Receive(serviceResponse, nil)
+	if err != nil {
 		return nil, errors.Wrap(err, "could not create service")
 	}
 
-	return res, nil
+	err = checkResponseStatusCode(response)
+	if err != nil {
+		return nil, err
+	}
+
+	return serviceResponse, nil
 }
 
 // UpdateService updates a old spankin service in FireHydrant
-// TODO: Check failure case
 func (c *RESTServicesClient) Update(ctx context.Context, serviceID string, updateReq UpdateServiceRequest) (*ServiceResponse, error) {
-	res := &ServiceResponse{}
-
-	if _, err := c.restClient().Patch("services/"+serviceID).BodyJSON(&updateReq).Receive(res, nil); err != nil {
+	serviceResponse := &ServiceResponse{}
+	response, err := c.restClient().Patch("services/"+serviceID).BodyJSON(&updateReq).Receive(serviceResponse, nil)
+	if err != nil {
 		return nil, errors.Wrap(err, "could not update service")
 	}
 
-	return res, nil
+	err = checkResponseStatusCode(response)
+	if err != nil {
+		return nil, err
+	}
+
+	return serviceResponse, nil
 }
 
 // DeleteService updates a old spankin service in FireHydrant
-// TODO: Check failure case
 func (c *RESTServicesClient) Delete(ctx context.Context, serviceID string) error {
-	if _, err := c.restClient().Delete("services/"+serviceID).Receive(nil, nil); err != nil {
+	response, err := c.restClient().Delete("services/"+serviceID).Receive(nil, nil)
+	if err != nil {
 		return errors.Wrap(err, "could not delete service")
+	}
+
+	err = checkResponseStatusCode(response)
+	if err != nil {
+		return err
 	}
 
 	return nil