+++ provider/functionality_resource.go
@@ -66,6 +66,11 @@ func readResourceFireHydrantFunctionality(ctx context.Context, d *schema.Resourc
 	// Get the functionality
 	r, err := firehydrantAPIClient.GetFunctionality(ctx, d.Id())
 	if err != nil {
+		_, isNotFoundError := err.(firehydrant.NotFound)
+		if isNotFoundError {
+			d.SetId("")
+			return nil
+		}
 		return diag.FromErr(err)
 	}
 
@@ -225,6 +230,11 @@ func deleteResourceFireHydrantFunctionality(ctx context.Context, d *schema.Resou
 	functionalityID := d.Id()
 	err := firehydrantAPIClient.DeleteFunctionality(ctx, functionalityID)
 	if err != nil {
+		_, isNotFoundError := err.(firehydrant.NotFound)
+		if isNotFoundError {
+			d.SetId("")
+			return nil
+		}
 		return diag.FromErr(err)
 	}